envi i386: fix i386SibOper getOperAddr range
Fix the i386SibOper.getOperAddr method that does not correctly compute the sum of
two registers that might be 'negative'. Essentially, mask the result to the pointer
size of the architecture.

This patch uses emu.imem_psize to fetch the architecture pointer size. Might not be
the correct method, but its a way I stumbled across.