keyPress event handler bugfix.
also, discovered the reason for having to explicitly handle all exceptions or crash:

http://pyqt.sourceforge.net/Docs/PyQt5/incompatibilities.html

as of PyQt5.5, PyQt no longer gracefully handles exceptions by printing a stack trace and continuing.  this is *designed* behavior.  i've handled as many exceptions as i know to, particularly in areas where users may extend Viv (including Interactive Python window execution).  i'm looking into a way to reverse this behavior, in case visi would prefer the old safety-net, but for the time-being, this is the path forward.  We have to explicitly handle our exceptions!