one option for fixing...  open to suggestions
add "codeflow" argument to getOperValue() for both Opcode and Operand.  VivWorkspace.makeOpcode() uses this option to avoid getting PC (or whatever PC's value is) when PC is referenced by opcodes such as ARM's "mov pc, lr".  this delta is intended to fix erroneous XREFs which are getting put in whenever PC is used as an operand.  due to ARM's pipelining, PC at the time of execution is +8, causing "mov pc, lr" to cause an XREF two instructions after the opcode, even though this instruction is IF_NOFALL.