setSymKid Speedup (#309)
A 40-60 percent reduction in runtime for symbolik reduction. Makes it so if the parent cache of a symbolik object is empty when we call setSymKid, we don't traverse all the way up the tree, due to some assumptions about how the caches get populated. See the setSymKid docstring in vivisect/symboliks/common.py for more information.