Keep Python2's behaviour of IMemory#getMemoryMap()
When using Python3, `getMemoryMap(None)` raises an exception, while with
Python2 it returns `None`. This is caused by the following breaking
change in Python3:
> The ordering comparison operators (<, <=, >=, >) raise a TypeError
> exception when the operands don’t have a meaningful natural ordering.

This causes `isValidPointer(None)` to also raise an exception in
Python3, while it returns `False` in Python2.

This can be a breaking change for applications relaying on being able to
call these methods with `None` as argument. For example (from capa):
```
while True:
    if not vw.isValidPointer(p):
        return
    yield p

    try:
        next = vw.readMemoryPtr(p)
```
https://github.com/fireeye/capa/blob/master/capa/features/extractors/viv/insn.py#L193