elf: reduce logging verbosity for status messages
reducing the logging level for "normal" messages during ELF parsing routines. i'd recommend that debug be used for expected, verbose messages, info be used for things a human really needs to know, and warning for errors that are handled without crashing the program. therefore, the fact that an ELF parsing routine is invoked or that a data structure is not found (but this is expected) should all be at the debug level.

as always, open to discussion.