making architecture names/numbers in envi for *impending* architectures. more to come after this.
we should be moving to a model where we allocate the name/number early in the arch dev process so we aren't constantly conflicting when merging things in.  merging a new architecture can take a bit of time, and maintaining different numbering/naming systems between branches is difficult, and ruins workspaces we generate in that time.

we'll soon also be working in a completeness and maturity rating for each arch.
completeness, meaning disasm, emu, symboliks, unittests coverage
maturity meaning alpha, beta, production, stable